refactor: put back map_msg to see the flow of mapping the msg between Node, Element, and Attributes
