+++ crates/core/src/vdom/attribute/attribute_value.rs
@@ -135,20 +135,4 @@ impl<MSG> AttributeValue<MSG> {
         matches!(self, Self::Empty)
     }
 
-    /// map the msg of this AttributeValue such that `AttributeValue<MSG>` becomes
-    /// `AttributeValue<MSG2>`
-    pub fn map_msg<F, MSG2>(self, cb: F) -> AttributeValue<MSG2>
-    where
-        F: Fn(MSG) -> MSG2 + Clone + 'static,
-        MSG2: 'static,
-        MSG: 'static,
-    {
-        match self {
-            AttributeValue::FunctionCall(this) => AttributeValue::FunctionCall(this),
-            AttributeValue::Simple(this) => AttributeValue::Simple(this),
-            AttributeValue::Style(this) => AttributeValue::Style(this),
-            AttributeValue::EventListener(this) => AttributeValue::EventListener(this.map_msg(cb)),
-            AttributeValue::Empty => AttributeValue::Empty,
-        }
-    }
 }
