+++ crates/core/src/vdom/node.rs
@@ -321,30 +321,6 @@ impl<MSG> Node<MSG> {
             .and_then(|att_values| att_values.first().and_then(|v| v.get_simple()))
     }
 
-    /// map the msg of this node such that Node<MSG> becomes Node<MSG2>
-    pub fn map_msg<F, MSG2>(self, cb: F) -> Node<MSG2>
-    where
-        F: Fn(MSG) -> MSG2 + Clone + 'static,
-        MSG2: 'static,
-        MSG: 'static,
-    {
-        match self {
-            Node::Element(element) => Node::Element(element.map_msg(cb)),
-            Node::Leaf(leaf) => Node::Leaf(leaf.map_msg(cb)),
-            Node::Fragment(nodes) => Node::Fragment(
-                nodes
-                    .into_iter()
-                    .map(|node| node.map_msg(cb.clone()))
-                    .collect(),
-            ),
-            Node::NodeList(node_list) => Node::NodeList(
-                node_list
-                    .into_iter()
-                    .map(|node| node.map_msg(cb.clone()))
-                    .collect(),
-            ),
-        }
-    }
 
     pub(crate) fn get_callbacks(&self) -> Vec<&EventCallback<MSG>> {
         if let Some(attributes) = self.attributes() {
